pyproject.toml
[project]
name = "front-office-pro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "motor>=3.7.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.17.1",
    "uvicorn[standard]>=0.42.0",
    "minio>=7.2.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pymongo>=4.0.0",
    "fastembed>=0.4.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
